deploy/requirements.txt
fabric==3.1.0
python-dotenv==1.0.1
colorama==0.4.6
GitPython==3.1.44